+++ src/patterns/SaleCard/SaleCard.tsx
@@ -1,12 +1,13 @@
 import { ComponentProps, forwardRef } from 'react';
 import { getCommonProps } from '../../utils';
 import classnames from 'classnames';
-import { Text, TextVariants } from '../Text';
-import { SeldonImage } from '../SeldonImage';
-import Button from '../Button/Button';
-import { ButtonVariants } from '../Button/types';
+import { Text, TextVariants } from '../../components/Text';
+import { SeldonImage } from '../../components/SeldonImage';
+import Button from '../../components/Button/Button';
+import { ButtonVariants } from '../../components/Button/types';
 import { SSRMediaQuery } from '../../providers/SeldonProvider/utils';
 import { SaleCardVariants } from './types';
+import { SaleCardActions } from './SaleCardActions';
 
 /** * Props for the SaleCard component. */
 export interface SaleCardProps extends ComponentProps<'div'> {
@@ -22,15 +23,6 @@ export interface SaleCardProps extends ComponentProps<'div'> {
   date: string;
   /** * The location of the sale. */
   location: string;
-  /** * The text to be displayed on the primary button. If not provided, the button will not be rendered. */
-  primaryButtonText?: string;
-  /** * The callback function to be executed when the primary button is clicked. Required if primaryButtonText is provided. */
-  primaryButtonOnClick?: () => void;
-  /** * The text to be displayed on the secondary button. If not provided, the button will not be rendered. */
-  secondaryButtonText?: string;
-  /** * The URL or callback function for the secondary button. Required if secondaryButtonText is provided. */
-  secondaryButtonHref?: string;
-  secondaryButtonOnClick?: () => void;
   /** * The text to be displayed as a badge. */
   badgeText?: string;
   /** * The text to be displayed for the modal link. */
@@ -42,6 +34,8 @@ export interface SaleCardProps extends ComponentProps<'div'> {
    * - 'relatedSaleTile': A variant with a smaller image size and horizontal layout on mobile devices.
    */
   variant?: SaleCardVariants;
+  /** * The <SaleCardActions /> component used to render the SaleCard CTAs. */
+  children?: React.ReactElement<typeof SaleCardActions>;
 }
 
 /**
@@ -51,7 +45,7 @@ export interface SaleCardProps extends ComponentProps<'div'> {
  *
  * [Figma Link](https://www.figma.com/design/hMu9IWH5N3KamJy8tLFdyV/Design-System--Responsive-Web?node-id=25028-2188&m=dev)
  *
- * [Storybook Link](https://phillips-seldon.netlify.app/?path=/docs/components-salecard--overview)
+ * [Storybook Link](https://phillips-seldon.netlify.app/?path=/docs/patterns-salecard--overview)
  */
 
 const SaleCard = forwardRef<HTMLDivElement, SaleCardProps>(
@@ -64,15 +58,11 @@ const SaleCard = forwardRef<HTMLDivElement, SaleCardProps>(
       titleText,
       date,
       location,
-      primaryButtonText,
-      primaryButtonOnClick,
-      secondaryButtonText,
-      secondaryButtonHref,
-      secondaryButtonOnClick,
       badgeText,
       modalButtonOnClick,
       modalButtonText,
       variant = SaleCardVariants.DEFAULT,
+      children,
       ...props
     },
     ref,
@@ -84,7 +74,7 @@ const SaleCard = forwardRef<HTMLDivElement, SaleCardProps>(
     const componentProps = { ...commonProps, ...props };
 
     return (
-      <div {...componentProps} className={classes} ref={ref}>
+      <article {...componentProps} className={classes} ref={ref}>
         {imageSrc ? <SeldonImage src={imageSrc} alt={imageAlt} className={`${baseClassName}__image`} /> : null}
         <div className={`${baseClassName}__details`}>
           <Text variant={TextVariants.badge} className={`${baseClassName}__auction-type`}>
@@ -113,34 +103,9 @@ const SaleCard = forwardRef<HTMLDivElement, SaleCardProps>(
           </div>
         </div>
         {variant !== SaleCardVariants.RELATED_SALE_TILE && (
-          <SSRMediaQuery.Media greaterThanOrEqual="md">
-            <div className={`${baseClassName}__cta`}>
-              {primaryButtonText && primaryButtonOnClick && (
-                <Button
-                  variant={ButtonVariants.primary}
-                  onClick={primaryButtonOnClick}
-                  className={`${baseClassName}__cta-button`}
-                >
-                  {primaryButtonText}
-                </Button>
-              )}
-              {secondaryButtonText && (
-                <Button
-                  variant={secondaryButtonHref ? ButtonVariants.tertiary : ButtonVariants.secondary}
-                  href={secondaryButtonHref}
-                  target={secondaryButtonHref ? '_blank' : undefined}
-                  onClick={secondaryButtonOnClick}
-                  className={classnames(`${baseClassName}__cta-button`, {
-                    [`${baseClassName}__pdf-link`]: secondaryButtonHref,
-                  })}
-                >
-                  {secondaryButtonText}
-                </Button>
-              )}
-            </div>
-          </SSRMediaQuery.Media>
+          <SSRMediaQuery.Media greaterThanOrEqual="md">{children}</SSRMediaQuery.Media>
         )}
-      </div>
+      </article>
     );
   },
 );