+++ CHANGELOG.md
@@ -1,3 +1,67 @@
+# [1.167.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.166.0...v1.167.0) (2025-07-22)
+
+
+### Features
+
+* **ExitgateCalendar:** L3-6952 Update image to match figma ([#665](https://github.com/PhillipsAuctionHouse/seldon/issues/665)) ([538f8d8](https://github.com/PhillipsAuctionHouse/seldon/commit/538f8d8c5974b05630b2e7463775cb9472c748eb))
+
+# [1.166.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.165.0...v1.166.0) (2025-07-21)
+
+
+### Features
+
+* **Viewing Details:** L3-6707 Add missing import to index ([#664](https://github.com/PhillipsAuctionHouse/seldon/issues/664)) ([0c343f6](https://github.com/PhillipsAuctionHouse/seldon/commit/0c343f6c0492c313b88d773424d1f650b1438b29))
+
+# [1.165.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.164.2...v1.165.0) (2025-07-21)
+
+
+### Bug Fixes
+
+* **ExitGateCard:** L3-6952 Remove text-transform capitalize for title ([#662](https://github.com/PhillipsAuctionHouse/seldon/issues/662)) ([312e853](https://github.com/PhillipsAuctionHouse/seldon/commit/312e8536a0cf20f710291ce7821ae32f1aa101fc))
+
+
+### Features
+
+* **_navigation:** L3-6501 header hover delay ([#640](https://github.com/PhillipsAuctionHouse/seldon/issues/640)) ([2952d75](https://github.com/PhillipsAuctionHouse/seldon/commit/2952d7543ca9ce1c7465a0bdf2e04fc85ec5f8ef))
+
+## [1.164.2](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.164.1...v1.164.2) (2025-07-21)
+
+
+### Bug Fixes
+
+* **icon:** L3-7333 Revert set width correctly for safari ([#661](https://github.com/PhillipsAuctionHouse/seldon/issues/661)) ([#663](https://github.com/PhillipsAuctionHouse/seldon/issues/663)) ([8e52f0c](https://github.com/PhillipsAuctionHouse/seldon/commit/8e52f0c59825b755840b80e7689b6853b52eb0fa))
+
+## [1.164.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.164.0...v1.164.1) (2025-07-21)
+
+
+### Bug Fixes
+
+* **icon:** L3-7333 set width correctly for safari ([#661](https://github.com/PhillipsAuctionHouse/seldon/issues/661)) ([1f54c0e](https://github.com/PhillipsAuctionHouse/seldon/commit/1f54c0ea8f139748f23ef8102ec0c3e326fdb15e))
+
+# [1.164.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.163.0...v1.164.0) (2025-07-17)
+
+
+### Features
+
+* **SaleCard:** L3-7305 update SaleCard to use SaleCardActions for CTAs ([#658](https://github.com/PhillipsAuctionHouse/seldon/issues/658)) ([2027efa](https://github.com/PhillipsAuctionHouse/seldon/commit/2027efa648fe1f94a62b870839aadfbadbb3ff34))
+
+# [1.163.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.162.0...v1.163.0) (2025-07-16)
+
+
+### Features
+
+* **ExitgateCard:** L3-6952 Add missing import to index ([#657](https://github.com/PhillipsAuctionHouse/seldon/issues/657)) ([9df0abf](https://github.com/PhillipsAuctionHouse/seldon/commit/9df0abf46550399510c6f449042f55fdd0d0ca38))
+
+# [1.162.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.161.0...v1.162.0) (2025-07-16)
+
+
+### Features
+
+* **Article:** L3-6952 Create exit gate card ([#639](https://github.com/PhillipsAuctionHouse/seldon/issues/639)) ([0d6eee0](https://github.com/PhillipsAuctionHouse/seldon/commit/0d6eee0d3e1585dcd27d3e1f099ff45305122573))
+* **ComposeModal:** L3-6977 Seldon: ViewingDetails Card + ComposedModal update ([#642](https://github.com/PhillipsAuctionHouse/seldon/issues/642)) ([56f8b28](https://github.com/PhillipsAuctionHouse/seldon/commit/56f8b287bbc45e7cda2309cdcee1b8f594366a02))
+* **iconButton:** L3-7252 fix tertiary hover ([#654](https://github.com/PhillipsAuctionHouse/seldon/issues/654)) ([2dd708c](https://github.com/PhillipsAuctionHouse/seldon/commit/2dd708c353a0cd89b78e3a861bc51a6b8a2a5ffc))
+* **RelatedSale:** L3-7218 Update auctio type font size ([#655](https://github.com/PhillipsAuctionHouse/seldon/issues/655)) ([fc968c5](https://github.com/PhillipsAuctionHouse/seldon/commit/fc968c5c8ada8f2874ae9d0f371305c99678d84f))
+
 # [1.161.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.160.0...v1.161.0) (2025-07-02)